bug fixes no delete por causa de associaçao

+++ makeaburguer/backend/controllers/PromocoesController.php
@@ -4,6 +4,7 @@ namespace backend\controllers;
 
 use Yii;
 use app\models\Promocoes;
+use app\models\Pedido;
 use backend\models\PromocoesSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
@@ -149,9 +150,22 @@ class PromocoesController extends Controller
     public function actionDelete($id)
     {
         if(Yii::$app->user->can('view-admin')) {
-            $this->findModel($id)->delete();
+            //$this->findModel($id)->delete();
 
-            return $this->redirect(['index']);
+            //return $this->redirect(['index']);
+
+            $verifica = Pedido::find()
+                ->select('id')
+                ->where(['promocao' => $id])
+                ->one();
+            if( $verifica !== null ) {
+
+                return $this->redirect(['pedido/view','id' => $verifica->id]);
+            }   
+            else{
+                $this->findModel($id)->delete();
+                return $this->redirect(['index']);
+            }
         }
         else{
             throw new ForbiddenHttpException();
